% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_orderbook_bittrex}
\alias{get_orderbook_bittrex}
\title{Get orderbook (bittrex)}
\usage{
get_orderbook_bittrex(level = 5)
}
\arguments{
\item{level}{Required orderbook level (default = 5, upper bound = 25)}
}
\value{
List with Timestamp (Unix-format), Ask Side (Price and Quantity), Bid (Price and Quantity)
}
\description{
This function scraps data from the orderbook API of Kraken
}
