% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_orderbook.R
\name{store_orderbook}
\alias{store_orderbook}
\title{Live Scrapper (whenever this file is running it gets the current orderbook snapshot and stores it)}
\usage{
store_orderbook(l = get_orderbook(), folder = "bitfinex")
}
\description{
Live Scrapper (whenever this file is running it gets the current orderbook snapshot and stores it)
}
